services/product/v1.0.1/pyproject.toml
[tool.poetry]
name = "product"
version = "1.0.1"
description = "product attributes / customization options for merchants"
authors = ["T.H."]
license = "MIT"
packages = [{include="product", from="./src"}, {include="settings", from="."}, {include="tests", from="."}]

[tool.poetry.dependencies]
python = "^3.12"
celery = "^5.4.0"
daphne = "3.0.1"
Django = "3.1.8" # TODO, upgrade to 5.0 , or switch to FastAPI + SQLAlchemy
djangorestframework = "3.11.2"
elasticsearch = "5.5.3"
ijson = "3.1.4"
mock = "4.0.3"
mysqlclient = "1.4.6"
Pillow = "7.2.0"
PyJWT = "2.8.0"
python-logstash-async = "2.5.0"
requests = "2.25.0"
ecommerce-common = {path = "../../common/python"}
my-c-extension-lib = {path = "../../common/python/c_exts"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"
black = "^24.4.0"

[tool.ruff]
exclude = ["./src/product/migrations/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
